[Refactor] #4971 calc_score() 周りから%ld のフォーマット指定子をなくした

+++ src/core/scores.cpp
@@ -189,7 +189,7 @@ errr top_twenty(PlayerType *player_ptr)
 {
     high_score the_score = {};
     snprintf(the_score.what, sizeof(the_score.what), "%u.%u.%u", H_VER_MAJOR, H_VER_MINOR, H_VER_PATCH);
-    snprintf(the_score.pts, sizeof(the_score.pts), "%9ld", (long)calc_score(player_ptr));
+    snprintf(the_score.pts, sizeof(the_score.pts), "%9u", calc_score(player_ptr));
     the_score.pts[9] = '\0';
 
     snprintf(the_score.gold, sizeof(the_score.gold), "%9lu", (long)player_ptr->au);
